refactor : 매칭기능 입력에 대한 비즈니스 로직을 서비스로 이동
