+++ src/main/java/pairmatching/controller/MatchingController.java
@@ -3,6 +3,7 @@ package pairmatching.controller;
 import pairmatching.domain.Course;
 import pairmatching.domain.Level;
 import pairmatching.domain.Missions;
+import pairmatching.service.MatchingSercive;
 
 import static pairmatching.utils.ExceptionMessage.*;
 import static pairmatching.utils.Validator.*;
@@ -12,6 +13,7 @@ import static pairmatching.view.OutputViews.printMissionInfo;
 
 public class MatchingController {
     public static final Missions missions = new Missions();
+    static final MatchingSercive matchingService = new MatchingSercive();
 
     public void run() {
         boolean flag = false;
@@ -19,7 +21,7 @@ public class MatchingController {
             try {
                 printMissionInfo(missions);
                 String input = getMatchingInfo();
-                checkMissionInfo(input);
+                matchingService.checkMissionInfo(input, missions);
                 System.out.println("성공");
                 flag = true;
             } catch (IllegalArgumentException e) {
@@ -28,32 +30,5 @@ public class MatchingController {
         }
     }
 
-    private void checkMissionInfo(String input) {
-        String[] infoArr = input.split(", ");
-        checkInfoCount(infoArr);
-        checkCourse(infoArr[0]);
-        checkLevel(infoArr[1]);
-        checkMission(infoArr[1], infoArr[2]);
-    }
 
-    private void checkCourse(String course) {
-        checkEachInfo(course);
-        if (!Course.isExist(course)) {
-            throw new IllegalArgumentException(ERROR_COURSE_NOT_EXIST);
-        }
-    }
-
-    private void checkLevel(String level) {
-        checkEachInfo(level);
-        if (!Level.isExist(level)) {
-            throw new IllegalArgumentException(ERROR_LEVEL_NOT_EXIST);
-        }
-    }
-
-    private void checkMission(String level, String mission) {
-        checkEachInfo(mission);
-        if (!missions.isExist(level, mission)) {
-            throw new IllegalArgumentException(ERROR_MISSION_NOT_EXIST);
-        }
-    }
 }
